requirements.txt
absl-py==1.4.0
agora-token-builder==1.0.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
dj-database-url==2.1.0
Django==4.0.1
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.12.4
flatbuffers==23.5.26
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.1
Pillow==10.0.0
platformdirs==3.11.0
protobuf==4.24.3
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
# tensorflow==2.13.0
# tensorflow-estimator==2.13.0
# tensorflow-intel==2.13.0
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.3.0
typing_extensions==4.5.0
tzdata==2024.1
urllib3==1.26.16
virtualenv==20.24.5
Werkzeug==2.3.7
whitenoise==6.6.0
wrapt==1.15.0
